requirements.txt
annotated-types==0.7.0
asttokens==2.4.1
asyncio==3.4.3
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
curl_cffi==0.6.4
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.20.0
gevent==24.2.1
greenlet==3.0.3
idna==3.7
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
MouseInfo==0.1.3
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
pure-eval==0.2.2
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytweening==1.2.0
pyzmq==26.0.3
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
setuptools==70.0.0
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
websocket==0.2.1
websockets==12.0
wheel==0.43.0
yarg==0.1.9
zope.event==5.0
characterai @ git+https://github.com/kramcat/CharacterAI.git@80b1bbff11859f6d0ec0510abb7f9e0982babb93